Implemented SentenceAutocompleteValidator form validator and is now used by ArgumentFixerDialog (refs #4)
It gets and expression and calls the dialogLauncher's autocomplete on it with all current values of the form.
It also needs the name of the current field that is validated, so BaseField now has a getter for that.
ArgumentFixerDialog now run autoValidate on all its fields after render, instead of manually adding the error message.
ArgumentFixerDialog now has some refactor on how to get the names of arguments it should deal with.
Also now adds an increasing index to the form name, since it is used as a prefix in the field id, wich lead to multiple inputs in several forms having the same id!