Consoloid.Ui.Dialog: It nowhas a startWithoutExpression method (refs #6)
Refactored start() and handleArguents(args, expression) to match this.
render() will not atempt to render the request if there is no expression to render.

Consoloid.Ui.Console#startWithDialog() now calls this, so it will be useful for welcome dialogs.

Consoloid.Ui.Volatile.Dialog: It used to have a handleArguments, that does exactly what it does now in Consoloid.Ui.Dialog.
Consoloid.Topic.ListDialog: Has three different start versions, they now all call startWithoutExpression in the end instead of manually calling setup and render (and not calling console to create the dialog's node)