Prompt: decrease autocomplete idleTimeout to eliminate unintended interpretations
Some string arguments may come up with highest rank in autocomplete and may
result in unintended interpretation of the sentence.

E.g. in tada "sw br ma in console" was interpreted to "create and switch to branch
ma in console". However I've meant "switch to branch master in console".

This quick workaround puts autocomplete on the screen soon enough for the
user the realize what's going to happen on pressing enter.